rename : move dir keyword and topic
